Set up boiler plater 3D space.
Created an 'App' export class and intialized all nescessary variables
and methods. Set up the renderer, scene and perspective camera, as well
as the raf fucntion and a step function for further use.

The raf method calls the requestAnimationFrame method. The step function
calculates the time elapsed since the last frame for further potetial
use.